Fixes deployment routes diff
