+++ packages/sdk/src/mcp/hosting/fs-polyfill.ts
@@ -1,4 +1,4 @@
-import type { Polyfill } from "./api.ts";
+import type { Polyfill } from "./deployment.ts";
 
 const promises = {
   stat(_path: string) {
